New variable names

+++ aggregator.js
@@ -1,19 +1,23 @@
 const fs = require('fs')
 
 function collect(result){
-    let data = fs.readFileSync('theResult.json');
-    data = JSON.parse(data);
-    Object.keys(result).forEach(name => {
-        result[name].forEach(offsets => {
-            if(data[name]){
-                data[name].push(offsets);
-            }else{
-                data[name] = [offsets];
-            }
+    if(!fs.existsSync('theResult.json')){
+        fs.writeFileSync('theResult.json', '{}');
+    }else{
+        let data = fs.readFileSync('theResult.json');
+        data = JSON.parse(data);
+        Object.keys(result).forEach(name => {
+            result[name].forEach(offsets => {
+                if(data[name]){
+                    data[name].push(offsets);
+                }else{
+                    data[name] = [offsets];
+                }
+            });
         });
-    });
-
-    fs.writeFileSync('theResult.json',JSON.stringify(data,null,2));
+        
+        fs.writeFileSync('theResult.json',JSON.stringify(data,null,2));
+    } 
 }
